Add the related logic to appReducer
