+++ src/context/products.js
@@ -4,71 +4,83 @@ export const products = [
       url: 'images/playstation_5.png',
       category: 'games',
       price: 499.99,
+      id: 21
     },
     {
       name: 'Samsung Galaxy',
       url: 'images/samsung_galaxy.png',
       category: 'smartphones',
       price: 399.99,
+      id: 22
     },
     {
       name: 'Cannon EOS Camera',
       url: 'images/cannon_eos_camera.png',
       category: 'cameras',
       price: 749.99,
+      id: 23
     },
     {
       name: 'Sony A7 Camera',
       url: 'images/sony_a7_camera.png',
       category: 'cameras',
       price: 1999.99,
+      id: 24
     },
     {
       name: 'LG TV',
       url: 'images/lg_tv.png',
       category: 'televisions',
       price: 799.99,
+      id: 456456
     },
     {
       name: 'Nintendo Switch',
       url: 'images/nintendo_switch.png',
       category: 'games',
       price: 299.99,
+      id: 351
     },
     {
       name: 'Xbox Series X',
       url: 'images/xbox_series_x.png',
       category: 'games',
       price: 499.99,
+      id: 25
     },
     {
       name: 'Samsung TV',
       url: 'images/samsung_tv.png',
       category: 'televisions',
       price: 1099.99,
+      id: 26
     },
     {
       name: 'Google Pixel',
       url: 'images/google_pixel.png',
       category: 'smartphones',
       price: 499.99,
+      id: 27
     },
     {
       name: 'Sony ZV1F Camera',
       url: 'images/sony_zv1f_camera.png',
       category: 'cameras',
       price: 799.99,
+      id: 28
     },
     {
       name: 'Toshiba TV',
       url: 'images/toshiba_tv.png',
       category: 'televisions',
       price: 499.99,
+      id: 29
     },
     {
       name: 'iPhone 14',
       url: 'images/iphone_14.png',
       category: 'smartphones',
       price: 999.99,
+      id: 30
     },
   ];
