+++ src/main.jsx
@@ -1,10 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
+import { AppProvider } from './context/globalState.js'
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    {/* App component needs to be wrapped by the AppProvider in order to have access to global state*/}
+    <AppProvider>
+      <App />
+    </AppProvider>
   </StrictMode>,
 )
